@@ -28,7 +28,7 @@
     "test-cov": "rm -rf coverage; nyc --reporter=html mocha"
   },
   "devDependencies": {
-    "@iabtechlabtcf/testing": "1.5.7",
+    "@iabtechlabtcf/testing": "1.5.8",
     "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "@types/sinon": "^10.0.11",
     "@types/sinon-chai": "3.2.8",
@@ -50,9 +50,9 @@
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.12",
     "terser": "^4.8.1",
-    "ts-node": "^10.5.0",
-    "typedoc": "^0.22.11",
-    "typescript": "4.5.5"
+    "ts-node": "10.9.1",
+    "typedoc": "0.24.8",
+    "typescript": "4.9.5"
   },
   "eslintIgnore": [
     "lib"
