@@ -41,8 +41,8 @@
     "mocha": "^9.2.0",
     "source-map-support": "^0.5.12",
     "terser": "^4.8.1",
-    "ts-node": "^10.5.0",
-    "typescript": "4.5.5"
+    "ts-node": "10.9.1",
+    "typescript": "4.9.5"
   },
   "eslintIgnore": [
     "lib"
